스코어가 제대로 표기 안되던 문제 해결

+++ Assets/Scripts/UI/UIMain.cs
@@ -16,7 +16,7 @@ public class UIMain : MonoBehaviour
     {
         MusicSlider.value = GameSystem.Instance.GetPlayerPref(PlayerPrefType.Music);
         SoundSlider.value = GameSystem.Instance.GetPlayerPref(PlayerPrefType.Sound);
-        HighScore.text = GameSystem.Instance.GetPlayerPref(PlayerPrefType.HighScore).ToString();
+        HighScore.text = ((int)GameSystem.Instance.GetPlayerPref(PlayerPrefType.HighScore)).ToString();
     }
 
     public void LateUpdate()
